+++ lib/screen/fss/document_delivery/form_document_delivery/form_document_delivery_screen.dart
@@ -321,7 +321,7 @@ class FormDocumentDeliveryScreen extends StatelessWidget {
                                   value: controller.receiverCompanyID.toString(),
                                   onChanged: (value) {
                                     controller.receiverCompanyID = value!.toInt();
-                                    controller.fetchLocationList(value.toInt());
+                                    controller.fetchLocationList(value);
                                     controller.update();
                                   },
                                 ),
